Create even number
In this program, a for loop iterates through numbers from 1 to 100. Within the loop, each number is checked to determine if it's even (i.e., divisible by 2 with no remainder). If the number is even, it's added to the sum variable. Finally, the program prints the sum of all even numbers from 1 to 100.