Update weather
Before running this code, make sure you have the libcurl and jsoncpp libraries installed on your system. You'll also need to sign up for an API key from OpenWeatherMap and replace "YOUR_API_KEY" with your actual API key.

This program prompts the user to enter the name of a city. It then sends a request to the OpenWeatherMap API to fetch weather data for that city. After receiving the response, it parses the JSON data and extracts the weather description and temperature, which it then displays to the user.